krux-pipeline/requirements.txt
anthropic>=0.49.0
openai>=1.65.0
supabase>=2.13.0
python-dotenv>=1.0.0
